third_party/blink/web_tests/media/media-fragments/TC0036-expected.txt


Title: TC0036
Fragment: 't=%33'
Comment: The media is requested from 3 seconds to the end. Note: %33 is equivalent to 3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
